backend-azure_functions/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
azure-functions==1.7.2
azure-functions-durable==1.1.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
furl==2.1.3
idna==3.2
multidict==5.2.0
orderedmultidict==1.0.1
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0
